feat: 공지사항 기능 구현 완료 및 코드 구조 정리
